キャッシュされたパスの表示コマンドを実装
