+++ src/script/pathfinder.ts
@@ -1,3 +1,4 @@
+import { roomLink } from "utility/log"
 import { ResultFailed, ResultSucceeded, ResultType } from "utility/result"
 
 export function findPath(startObjectId: string, goalObjectId: string, goalRange: number): string {
@@ -49,7 +50,7 @@ interface SourceRoute {
   harvestPositions: RoomPosition[]
 }
 
-// TODO: Sourceが埋まっていたらよくないことが起きる
+// TODO: Sourceが壁等で埋まっていたらよくないことが起きる
 export function calculateSourceRoute(sourceId: Id<Source>, destination: RoomPosition): ResultType<SourceRoute, string> {
   const source = Game.getObjectById(sourceId)
   if (!(source instanceof Source)) {
@@ -103,4 +104,47 @@ export function calculateSourceRoute(sourceId: Id<Source>, destination: RoomPosi
     harvestPositions,
   }
   return new ResultSucceeded(result)
+}
+
+export function showCachedSourcePath(sourceId: Id<Source>): string {
+  const source = Game.getObjectById(sourceId)
+  if (source == null) {
+    return `Invalid source ID ${sourceId}`
+  }
+
+  const cachedPath = getCachedPathFor(source)
+  if (cachedPath == null) {
+    return `No cached source path for source ${sourceId} in ${roomLink(source.room.name)}`
+  }
+
+  const visual = source.room.visual
+  cachedPath.forEach(position => visual.text("*", position.x, position.y))
+  return "ok"
+}
+
+// Roomに持たせる
+const sourcePathCache = new Map<Id<Source>, RoomPosition[]>()
+
+export function getCachedPathFor(source: Source): RoomPosition[] | null {
+  const cachedPath = sourcePathCache.get(source.id)
+  if (cachedPath != null) {
+    if (cachedPath.length <= 0) {
+      return null
+    } else {
+      return cachedPath
+    }
+  }
+  if (source.room.memory.p == null) {
+    sourcePathCache.set(source.id, [])
+    return null
+  }
+  const memoryCachedPath = source.room.memory.p.s[source.id]
+  if (memoryCachedPath == null || memoryCachedPath === "no path") {
+    sourcePathCache.set(source.id, [])
+    return null
+  }
+  const roomName = source.room.name
+  const roomPositions = memoryCachedPath.p.map(position => new RoomPosition(position.x, position.y, roomName))
+  sourcePathCache.set(source.id, roomPositions)
+  return roomPositions
 }
